Move create_time_index to helpers
