+++ src/oemof/solph/__init__.py
@@ -8,8 +8,8 @@ from . import helpers
 from . import processing
 from . import views
 from ._energy_system import EnergySystem
-from ._energy_system import create_time_index
 from ._groupings import GROUPINGS
+from ._helpers import create_time_index
 from ._models import Model
 from ._options import Investment
 from ._options import NonConvex
