+++ src/oemof/solph/_energy_system.py
@@ -14,8 +14,6 @@ SPDX-License-Identifier: MIT
 
 """
 
-import calendar
-import datetime
 import warnings
 
 import numpy as np
@@ -244,71 +242,3 @@ class EnergySystem(es.EnergySystem):
             - self.periods[period].min().year
             + 1
         )
-
-
-def create_time_index(
-    year: int = None,
-    interval: float = 1,
-    number: int = None,
-    start: datetime.datetime or datetime.date = None,
-):
-    """
-    Create a datetime index for one year.
-
-    Notes
-    -----
-    To create 8760 hourly intervals for a non leap year a datetime index with
-    8761 time points need to be created. So the number of time steps is always
-    the number of intervals plus one.
-
-    Parameters
-    ----------
-    year : int, datetime
-        The year of the index. If number and start is set the year parameter is
-        ignored.
-    interval : float
-        The time interval in hours e.g. 0.5 for 30min or 2 for a two hour
-        interval (default: 1).
-    number : int
-        The number of time intervals. By default number is calculated to create
-        an index of one year. For a shorter or longer period the number of
-        intervals can be set by the user.
-    start : datetime.datetime or datetime.date
-        Optional start time. If start is not set, 00:00 of the first day of
-        the given year is the start time.
-
-    Examples
-    --------
-    >>> len(create_time_index(2014))
-    8761
-    >>> len(create_time_index(2012))  # leap year
-    8785
-    >>> len(create_time_index(2014, interval=0.5))
-    17521
-    >>> len(create_time_index(2014, interval=0.5, number=10))
-    11
-    >>> len(create_time_index(2014, number=10))
-    11
-    >>> str(create_time_index(2014, interval=0.5, number=10)[-1])
-    '2014-01-01 05:00:00'
-    >>> str(create_time_index(2014, interval=2, number=10)[-1])
-    '2014-01-01 20:00:00'
-    """
-    if number is None:
-        if calendar.isleap(year):
-            hoy = 8784
-        else:
-            hoy = 8760
-        number = round(hoy / interval)
-    if start is None:
-        start = f"1/1/{year}"
-    try:
-        time_index = pd.date_range(
-            start, periods=number + 1, freq=f"{interval}h"
-        )
-    except ValueError:
-        # Pandas <2.2 compatibility
-        time_index = pd.date_range(
-            start, periods=number + 1, freq=f"{interval}H"
-        )
-    return time_index
